backend/requirements.txt
Django>=5.0,<6
django-cors-headers>=4.3
python-dotenv>=1.0
Pillow>=10.0
